requirements.txt
Django==4.2
Pillow==9.5.0
psycopg2==2.9.6
django_cleanup
django_summernote
django-debug-toolbar~=4.1.0
django-redis
celery[redis]
djangorestframework
